fix(Task): fix task request validation

+++ project/app/Http/Controllers/TaskController.php
@@ -2,7 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Task as TaskRequest;
+use App\Http\Requests\TaskCreate as TaskCreateRequest;
+use App\Http\Requests\TaskUpdate as TaskUpdateRequest;
 use App\Task;
 
 class TaskController extends Controller
@@ -14,27 +15,45 @@ class TaskController extends Controller
         return response()->json($tasks);
     }
 
-    public function store(TaskRequest $request)
+    public function store(TaskCreateRequest $request)
     {
         $task = Task::create($request->all());
 
         return response()->json($task, 201);
     }
 
-    public function show(Task $task)
+    public function show($id)
     {
+        $task = Task::where('id', $id)->first();
+
+        if (!$task) {
+            return response(null, 404);
+        }
+
         return response()->json($task);
     }
 
-    public function update(TaskRequest $request, Task $task)
+    public function update(TaskUpdateRequest $request, $id)
     {
+        $task = Task::where('id', $id)->first();
+
+        if (!$task) {
+            return response(null, 404);
+        }
+
         $task->update($request->all());
 
         return response()->json($task, 202);
     }
 
-    public function destroy(Task $task)
+    public function destroy($id)
     {
+        $task = Task::where('id', $id)->first();
+
+        if (!$task) {
+            return response(null, 404);
+        }
+
         $task->delete();
 
         return response(null,204);
